# docs/openapi/climate-api-v1.json
{
    "openapi":  "3.0.1",
    "info":  {
                 "title":  "Climate Monitoring System API",
                 "description":  "Public API exposed by Climate.Gateway. Internal service endpoints are intentionally excluded.",
                 "version":  "v1"
             },
    "servers":  [
                    {
                        "url":  "http://localhost:8080",
                        "description":  "Local Docker Gateway"
                    }
                ],
    "paths":  {
...
                  "/api/monitoring/current":  {
                                                  "get":  {
                                                              "tags":  [
                                                                           "Monitoring"
                                                                       ],
                                                              "responses":  {
                                                                                "200":  {
                                                                                            "description":  "OK",
                                                                                            "content":  {
                                                                                                            "text/plain":  {
                                                                                                                               "schema":  {
                                                                                                                                              "type":  "array",
                                                                                                                                              "items":  {
                                                                                                                                                            "$ref":  "#/components/schemas/SensorReadingResponse"
                                                                                                                                                        }
                                                                                                                                          }
                                                                                                                           },
                                                                                                            "application/json":  {
                                                                                                                                     "schema":  {
                                                                                                                                                    "type":  "array",
                                                                                                                                                    "items":  {
                                                                                                                                                                  "$ref":  "#/components/schemas/SensorReadingResponse"
                                                                                                                                                              }
                                                                                                                                                }
                                                                                                                                 },
                                                                                                            "text/json":  {
                                                                                                                              "schema":  {
                                                                                                                                             "type":  "array",
                                                                                                                                             "items":  {
                                                                                                                                                           "$ref":  "#/components/schemas/SensorReadingResponse"
                                                                                                                                                       }
                                                                                                                                         }
                                                                                                                          }
                                                                                                        }
                                                                                        },
                                                                                "401":  {
                                                                                            "description":  "Unauthorized"
                                                                                        },
                                                                                "403":  {
                                                                                            "description":  "Forbidden"
                                                                                        }
                                                                            },
                                                              "operationId":  "Monitoring_get_api_monitoring_current",
                                                              "security":  [
                                                                               {
                                                                                   "Bearer":  [

                                                                                              ]
                                                                               }
                                                                           ]
                                                          }
                                              },
                  "/api/monitoring/sensors/{sensorId}/latest":  {
                                                                    "get":  {
                                                                                "tags":  [
                                                                                             "Monitoring"
                                                                                         ],
                                                                                "parameters":  [
                                                                                                   {
                                                                                                       "name":  "sensorId",
                                                                                                       "in":  "path",
                                                                                                       "required":  true,
                                                                                                       "schema":  {
                                                                                                                      "type":  "string",
                                                                                                                      "format":  "uuid"
                                                                                                                  }
                                                                                                   }
                                                                                               ],
                                                                                "responses":  {
                                                                                                  "200":  {
                                                                                                              "description":  "OK",
                                                                                                              "content":  {
                                                                                                                              "text/plain":  {
                                                                                                                                                 "schema":  {
                                                                                                                                                                "$ref":  "#/components/schemas/SensorReadingResponse"
                                                                                                                                                            }
                                                                                                                                             },
                                                                                                                              "application/json":  {
                                                                                                                                                       "schema":  {
                                                                                                                                                                      "$ref":  "#/components/schemas/SensorReadingResponse"
                                                                                                                                                                  }
                                                                                                                                                   },
                                                                                                                              "text/json":  {
                                                                                                                                                "schema":  {
                                                                                                                                                               "$ref":  "#/components/schemas/SensorReadingResponse"
                                                                                                                                                           }
                                                                                                                                            }
                                                                                                                          }
                                                                                                          },
                                                                                                  "401":  {
                                                                                                              "description":  "Unauthorized"
                                                                                                          },
                                                                                                  "403":  {
                                                                                                              "description":  "Forbidden"
                                                                                                          }
                                                                                              },
                                                                                "operationId":  "Monitoring_get_api_monitoring_sensors_sensorId_latest",
                                                                                "security":  [
                                                                                                 {
                                                                                                     "Bearer":  [

                                                                                                                ]
                                                                                                 }
                                                                                             ]
                                                                            }
                                                                },
                  "/api/monitoring/sensors/{sensorId}/history":  {
                                                                     "get":  {
                                                                                 "tags":  [
                                                                                              "Monitoring"
                                                                                          ],
                                                                                 "parameters":  [
                                                                                                    {
                                                                                                        "name":  "sensorId",
                                                                                                        "in":  "path",
                                                                                                        "required":  true,
                                                                                                        "schema":  {
                                                                                                                       "type":  "string",
                                                                                                                       "format":  "uuid"
                                                                                                                   }
                                                                                                    },
                                                                                                    {
                                                                                                        "name":  "communityId",
                                                                                                        "in":  "query",
                                                                                                        "schema":  {
                                                                                                                       "type":  "string",
                                                                                                                       "format":  "uuid"
                                                                                                                   }
                                                                                                    },
                                                                                                    {
                                                                                                        "name":  "from",
                                                                                                        "in":  "query",
                                                                                                        "schema":  {
                                                                                                                       "type":  "string",
                                                                                                                       "format":  "date-time"
                                                                                                                   }
                                                                                                    },
                                                                                                    {
                                                                                                        "name":  "to",
                                                                                                        "in":  "query",
                                                                                                        "schema":  {
                                                                                                                       "type":  "string",
                                                                                                                       "format":  "date-time"
                                                                                                                   }
                                                                                                    }
                                                                                                ],
                                                                                 "responses":  {
                                                                                                   "200":  {
                                                                                                               "description":  "OK",
                                                                                                               "content":  {
                                                                                                                               "text/plain":  {
                                                                                                                                                  "schema":  {
                                                                                                                                                                 "type":  "array",
                                                                                                                                                                 "items":  {
                                                                                                                                                                               "$ref":  "#/components/schemas/SensorReadingResponse"
                                                                                                                                                                           }
                                                                                                                                                             }
                                                                                                                                              },
                                                                                                                               "application/json":  {
                                                                                                                                                        "schema":  {
                                                                                                                                                                       "type":  "array",
                                                                                                                                                                       "items":  {
                                                                                                                                                                                     "$ref":  "#/components/schemas/SensorReadingResponse"
                                                                                                                                                                                 }
                                                                                                                                                                   }
                                                                                                                                                    },
                                                                                                                               "text/json":  {
                                                                                                                                                 "schema":  {
                                                                                                                                                                "type":  "array",
                                                                                                                                                                "items":  {
                                                                                                                                                                              "$ref":  "#/components/schemas/SensorReadingResponse"
                                                                                                                                                                          }
                                                                                                                                                            }
                                                                                                                                             }
                                                                                                                           }
                                                                                                           },
                                                                                                   "401":  {
                                                                                                               "description":  "Unauthorized"
                                                                                                           },
                                                                                                   "403":  {
                                                                                                               "description":  "Forbidden"
                                                                                                           }
                                                                                               },
                                                                                 "operationId":  "Monitoring_get_api_monitoring_sensors_sensorId_history",
                                                                                 "security":  [
                                                                                                  {
                                                                                                      "Bearer":  [

                                                                                                                 ]
                                                                                                  }
                                                                                              ]
                                                                             }
                                                                 },
...
              },
    "components":  {
                       "schemas":  {
...
                                       "SensorType":  {
                                                          "enum":  [
                                                                       "Temperature",
                                                                       "Humidity",
                                                                       "WindSpeed",
                                                                       "Rainfall",
                                                                       "WaterLevel"
                                                                   ],
                                                          "type":  "string"
                                                      },
...
                                       "SensorReadingResponse":  {
                                                                     "type":  "object",
                                                                     "properties":  {
                                                                                        "id":  {
                                                                                                   "type":  "string",
                                                                                                   "format":  "uuid"
                                                                                               },
                                                                                        "sensorId":  {
                                                                                                         "type":  "string",
                                                                                                         "format":  "uuid"
                                                                                                     },
                                                                                        "communityId":  {
                                                                                                            "type":  "string",
                                                                                                            "format":  "uuid"
                                                                                                        },
                                                                                        "sensorType":  {
                                                                                                           "$ref":  "#/components/schemas/SensorType"
                                                                                                       },
                                                                                        "value":  {
                                                                                                      "type":  "number",
                                                                                                      "format":  "double"
                                                                                                  },
                                                                                        "unit":  {
                                                                                                     "type":  "string",
                                                                                                     "nullable":  true
                                                                                                 },
                                                                                        "recordedAt":  {
                                                                                                           "type":  "string",
                                                                                                           "format":  "date-time"
                                                                                                       }
                                                                                    },
                                                                     "additionalProperties":  false
                                                                 },
...
                                   },
                       "securitySchemes":  {
                                               "Bearer":  {
                                                              "type":  "http",
                                                              "scheme":  "bearer",
                                                              "bearerFormat":  "JWT",
                                                              "description":  "Paste only the accessToken returned by POST /api/auth/login. Do not include the 'Bearer ' prefix; Swagger UI adds it automatically."
                                                          }
                                           }
                   },
...
}